refactor: split ToolRegistry into mixin classes
Extract 7 mixin classes from tool_registry.py (1459 → 454 lines):
- _callbacks.py: ChangeCallbackMixin
- _namespace.py: NamespaceMixin
- _enable_disable.py: EnableDisableMixin
- _registration.py: RegistrationMixin
- _permissions.py: PermissionsMixin
- _logging.py: ExecutionLoggingMixin
- _admin.py: AdminMixin

Public API unchanged. Closes #94.
